Advanced file sharing
This is a WIP for advanced file sharing, but I won't pursue this for now since using symlinks in the file system works for me, and this is adding a ton of complexity I don't want or need right now.

+++ routes/api/files/delete-directory.tsx
@@ -1,7 +1,7 @@
 import { Handlers } from 'fresh/server.ts';
 
 import { Directory, FreshContextState } from '/lib/types.ts';
-import { deleteDirectoryOrFile, getDirectories } from '/lib/data/files.ts';
+import { deleteDirectoryOrFile, getDirectories, getDirectoryAccess } from '/lib/data/files.ts';
 
 interface Data {}
 
@@ -30,11 +30,19 @@ export const handler: Handlers<Data, FreshContextState> = {
       return new Response('Bad Request', { status: 400 });
     }
 
-    // TODO: Verify user has write access to path and get the appropriate ownerUserId
-
-    const deletedDirectory = await deleteDirectoryOrFile(
+    const { hasWriteAccess, ownerUserId, ownerParentPath } = await getDirectoryAccess(
       context.state.user.id,
       requestBody.parentPath,
+      requestBody.name.trim(),
+    );
+
+    if (!hasWriteAccess) {
+      return new Response('Forbidden', { status: 403 });
+    }
+
+    const deletedDirectory = await deleteDirectoryOrFile(
+      ownerUserId,
+      ownerParentPath,
       requestBody.name.trim(),
     );
